#6 회원 가입 절반 성공 user_id 자동 증가 및 create_at sysdate 작동 불가 상태, 로그인도 불가 상태 아이디 중복 검사
