+++ canReform_table.sql
@@ -6,14 +6,6 @@ DROP TABLE ESTIMATES;
 DROP TABLE POSTS;
 DROP TABLE USERS;
 
-SELECT * FROM COMMENTS;
-SELECT * FROM RATINGS;
-SELECT * FROM REPORTS;
-SELECT * FROM SCHEDULE;
-SELECT * FROM ESTIMATES;
-SELECT * FROM POSTS;
-SELECT * FROM USERS;
-
 -- 1. USERS (회원)
 CREATE TABLE USERS (
     USER_ID NUMBER(10) PRIMARY KEY,
@@ -48,7 +40,7 @@ BEGIN
         VALUES (i, 'user' || i, 'user' || i || '@example.com', 'password' || i, SYSDATE);
     END LOOP;
 END;
-
+/
 -- 2. POSTS(게시판 글)
 CREATE TABLE POSTS (
     POST_ID NUMBER(10) PRIMARY KEY,
@@ -88,7 +80,7 @@ BEGIN
     END LOOP;
 END;
 
-
+/
 
 
 
@@ -156,3 +148,12 @@ CREATE TABLE ESTIMATES (
     FOREIGN KEY (POST_ID) REFERENCES POSTS(POST_ID),
     FOREIGN KEY (USER_ID) REFERENCES USERS(USER_ID)
 );
+
+
+SELECT * FROM COMMENTS;
+SELECT * FROM RATINGS;
+SELECT * FROM REPORTS;
+SELECT * FROM SCHEDULE;
+SELECT * FROM ESTIMATES;
+SELECT * FROM POSTS;
+SELECT * FROM USERS;
