+++ src/main/java/org/iclass/board/controller/UserController.java
@@ -13,12 +13,11 @@ import org.springframework.web.bind.annotation.*;
 
 @Controller
 @RequiredArgsConstructor
-@CrossOrigin(origins = "http://localhost:5173")
 @Slf4j
 public class UserController {
     private final UserService userService;
 
-/*    @GetMapping("/login")
+   @GetMapping("/login")
     public String showLoginForm() {
         return "login"; // login.html
     }
@@ -35,7 +34,7 @@ public class UserController {
 
         // 로그인 성공 후 처리 (예: 세션에 사용자 정보 저장)
         return "redirect:/home"; // 로그인 후 이동할 페이지
-    }*/
+    }
 
 
     @PostMapping("/signup")
@@ -51,6 +50,43 @@ public class UserController {
         session.invalidate();
         return "redirect:/login"; // 로그아웃 후 로그인 페이지로 리다이렉트
     }*/
+
+
+    /*    @GetMapping({"/", "/index"})
+    public String index(Model model, HttpServletRequest request, Authentication authentication) {
+
+        // 세션 ID 확인
+        HttpSession session = request.getSession();
+        log.info("로그인 후 세션 ID: {}", session.getId());
+
+        log.info("사용자 인증 : {}", authentication);
+        if(authentication != null && authentication.getPrincipal() instanceof UserDetails){
+            model.addAttribute("user", authentication.getName());
+        } else {
+            log.info("사용자 인증 정보 없음");
+        }
+        return "index";
+    }*/
+
+/*    @GetMapping("/login")
+    public String login(String error, Model model){
+        if(error != null && error.isEmpty()) {
+            model.addAttribute("error", "계정 정보가 일치하지 않습니다.");
+            log.info("error : {}", error);
+        }
+        return "login";
+    }*/
+
+    @GetMapping("/signup")
+    public String signup(){
+        return "signup";
+    }
+
+    @GetMapping("/posts")
+    public String bulletinBoard(){
+        return "posts";
+    }
+
 }
