+++ src/main/java/org/iclass/board/repository/UserRepository.java
@@ -6,6 +6,7 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface UserRepository extends JpaRepository<UserEntity, Long> {
+    UserEntity findByUsername(String username);
 
-
+    boolean existsByUsername(String username);
 }
